+++ lib/concurrent/atomic_reference/mutex_atomic.rb
@@ -1,4 +1,4 @@
-require 'thread'
+require 'concurrent/synchronization'
 require 'concurrent/atomic_reference/direct_update'
 require 'concurrent/atomic_reference/numeric_cas_wrapper'
 
@@ -8,31 +8,31 @@ module Concurrent
   #
   # @!visibility private
   # @!macro internal_implementation_note
-  class MutexAtomicReference
+  class MutexAtomicReference < Synchronization::Object
     include Concurrent::AtomicDirectUpdate
     include Concurrent::AtomicNumericCompareAndSetWrapper
 
     # @!macro atomic_reference_method_initialize
     def initialize(value = nil)
-      @mutex = Mutex.new
-      @value = value
+      super()
+      synchronize { ns_initialize(value) }
     end
 
     # @!macro atomic_reference_method_get
     def get
-      @mutex.synchronize { @value }
+      synchronize { @value }
     end
     alias_method :value, :get
 
     # @!macro atomic_reference_method_set
     def set(new_value)
-      @mutex.synchronize { @value = new_value }
+      synchronize { @value = new_value }
     end
     alias_method :value=, :set
 
     # @!macro atomic_reference_method_get_and_set
     def get_and_set(new_value)
-      @mutex.synchronize do
+      synchronize do
         old_value = @value
         @value = new_value
         old_value
@@ -42,7 +42,7 @@ module Concurrent
 
     # @!macro atomic_reference_method_compare_and_set
     def _compare_and_set(old_value, new_value)
-      @mutex.synchronize do
+      synchronize do
         if @value.equal? old_value
           @value = new_value
           true
@@ -50,6 +50,12 @@ module Concurrent
           false
         end
       end
+    end
+
+    protected
+
+    def ns_initialize(value)
+      @value = value
     end
   end
 end