+++ lib/concurrent/ivar.rb
@@ -1,5 +1,3 @@
-require 'thread'
-
 require 'concurrent/errors'
 require 'concurrent/collection/copy_on_write_observer_set'
 require 'concurrent/concern/obligation'