+++ spec/concurrent/executor/executor_options_spec.rb
@@ -29,53 +29,5 @@ module Concurrent
         Executor.executor_from_options(executor: :bogus)
       }.to raise_error(ArgumentError)
     end
-
-    it 'returns the global fast executor when :operation is true' do
-      expect(Concurrent).to receive(:global_fast_executor).
-                                and_return(:fast_executor)
-      Executor.executor_from_options(operation: true)
-    end
-
-    it 'returns the global io executor when :operation is false' do
-      expect(Concurrent).to receive(:global_io_executor).
-                                and_return(:io_executor)
-      Executor.executor_from_options(operation: false)
-    end
-
-    it 'returns the global fast executor when :task is false' do
-      expect(Concurrent).to receive(:global_fast_executor).
-                                and_return(:fast_executor)
-      Executor.executor_from_options(task: false)
-    end
-
-    it 'returns the global io executor when :task is true' do
-      expect(Concurrent).to receive(:global_io_executor).
-                                and_return(:io_executor)
-      Executor.executor_from_options(task: true)
-    end
-
-    it 'returns nil when :executor is nil' do
-      expect(Executor.executor_from_options(executor: nil)).to be_nil
-    end
-
-    it 'returns nil when no option is given' do
-      expect(Executor.executor_from_options).to be_nil
-    end
-
-    specify ':executor overrides :operation' do
-      expect(Executor.executor_from_options(executor: executor, operation: true)).
-          to eq executor
-    end
-
-    specify ':executor overrides :task' do
-      expect(Executor.executor_from_options(executor: executor, task: true)).
-          to eq executor
-    end
-
-    specify ':operation overrides :task' do
-      expect(Concurrent).to receive(:global_fast_executor).
-                                and_return(:fast_executor)
-      Executor.executor_from_options(operation: true, task: true)
-    end
   end
 end